# server/openapi.yaml
...
openapi: 3.1.0
info:
  title: Fake Stack Overflow API
  version: 1.0.0
  description: OpenAPI 3.1 specification for the Fake Stack Overflow backend server.
servers:
  - url: /api
    description: The API server endpoint
paths:
...
  /user/signup:
    post:
      tags:
        - Users
      summary: Create a new user account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSignupBody'
      responses:
        '200':
          description: User successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /user/login:
    post:
      tags:
        - Users
      summary: Log a user in.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginBody'
      responses:
        '200':
          description: User successfully logged in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /user/resetPassword:
    patch:
      tags:
        - Users
      summary: Reset a user's password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginBody'
      responses:
        '200':
          description: User password successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /user/getUser/{username}:
    get:
      tags:
        - Users
      summary: Retrieve a user by username.
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: The username of the user to retrieve.
      responses:
        '200':
          description: Successfully retrieved user details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /user/getUsers:
    get:
      tags:
        - Users
      summary: Retrieve a list of all users.
      responses:
        '200':
          description: Successfully retrieved list of all users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /user/deleteUser/{username}:
    delete:
      tags:
        - Users
      summary: Delete a user by username.
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: The username of the user to delete.
      responses:
        '200':
          description: User successfully deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /user/updateBiography:
    patch:
      tags:
        - Users
      summary: Update a user's biography.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBiographyBody'
      responses:
        '200':
          description: User biography successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /user/updateExternalLinks:
    patch:
      tags:
        - Users
      summary: Update a user's external links (LinkedIn, GitHub, Portfolio).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExternalLinksBody'
      responses:
        '200':
          description: User external links successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ErrorResponse'
components:
  schemas:
...
    UserSignupBody:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          nullable: true
          description: Optional user email.
        biography:
          type: string
          default: ''
          description: Optional user biography.
    UserLoginBody:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
    UpdateBiographyBody:
      type: object
      required:
        - username
        - biography
      properties:
        username:
          type: string
          minLength: 1
        biography:
          type: string
    UpdateExternalLinksBody:
      type: object
      required:
        - username
        - externalLinks
      properties:
        username:
          type: string
          minLength: 1
        externalLinks:
          type: object
          properties:
            linkedin:
              type: string
              format: uri
            github:
              type: string
              format: uri
            portfolio:
              type: string
              format: uri
    User:
      type: object
      description: A user profile object.
...
  responses:
...
    ErrorResponse:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: The error message returned by the server.
    ValidationError:
      description: Bad Request / Validation Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: The error message from express-openapi-validator.
              errors:
                type: array
                description: Detailed list of validation errors.